refactor(scripts): introduce barrel file to /puppeteer
