+++ scripts/puppeteer/index.ts
@@ -1,0 +1,1 @@
+export { safeLaunchOptions } from './puppeteer.config';
